[SOL-160] refactor: moved `test_public_withdraw_fail_with_unwhitelisted_resolver` to spl tests
